+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `coco_align` no longer errors and simply writes an empty kwcoco file if the input regions are empty by default. Old behavior can be enabled with 
 * Updated wrapper scoring code to support latest IARPA metrics.
 * Moved `prepare-ta2-dataset`, `prepare-splits`, and `prepare-teamfeats` to `cli/queue_cli`
+* The tracker now warns the user if it might overload memory
 
 ### Fixed
 * Cluster sites no longer includes system-rejected sites